+++ packages/fighting-design/components.ts
@@ -201,9 +201,6 @@ export * from './menu'
 export { FMenuItem } from './menu-item'
 export * from './menu-item'
 
-// export { FSlider } from './slider'
-// export * from './slider'
-
 export { FSubmenu } from './submenu'
 export * from './submenu'
 
@@ -212,3 +209,6 @@ export * from './menu-item-group'
 
 export { FCollapseAnimation } from './collapse-animation'
 export * from './collapse-animation'
+
+export { FSlider } from './slider'
+export * from './slider'