+++ start/src/App.vue
@@ -1,6 +1,8 @@
 <script lang="ts" setup>
   import { ref } from 'vue'
 
+  const value1 = ref(50)
+
   const current = ref(1)
   const total = ref(100)
 
@@ -14,7 +16,12 @@
 </script>
 
 <template>
-  <f-pagination v-model:current="current" round :total="total" />
+  <br />
+  <br />
+  <br />
+  <f-slider v-model="value1"></f-slider>
+
+  <!-- <f-pagination v-model:current="current" round :total="total" />
   <f-pagination
     v-model:current="current"
     round
@@ -22,5 +29,5 @@
     :total="total"
     :on-next="onNext"
     :on-prev="onPrev"
-  />
+  /> -->
 </template>